change signature of predict / learn task not in model because of memory usage. r sucks

+++ src/base/wrapped.model.r
@@ -27,7 +27,7 @@ setClass(
 		representation = representation(
 				learn.task = "learn.task",
 				learner.model = "ANY",
-				subset = "integer",
+				subset = "numeric",
 				parset = "list"
 		)
 )
